requirements.txt
spotipy>=2.19.0
syncedlyrics>=0.3.0
python-dotenv>=0.19.0